+++ src/main/java/calculator/Main.java
@@ -1,5 +1,14 @@
 package calculator;
 
+import calculator.numbers.AngleConverter;
+import calculator.numbers.MyNumber;
+import calculator.numbers.RationalNumber;
+import calculator.numbers.RealNumber;
+import calculator.operations.Divides;
+import calculator.operations.Minus;
+import calculator.operations.Plus;
+import calculator.operations.Times;
+
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
