+++ src/main/java/calculator/operations/Divides.java
@@ -1,4 +1,9 @@
-package calculator;
+package calculator.operations;
+
+import calculator.Expression;
+import calculator.IllegalConstruction;
+import calculator.Notation;
+import calculator.numbers.RationalNumber;
 
 import java.util.List;
 
@@ -18,7 +23,7 @@ public final class Divides extends Operation
    *
    * @param elist The list of Expressions to divide
    * @throws IllegalConstruction    If an empty list of expressions if passed as parameter
-   * @see #Divides(List<Expression>,Notation)
+   * @see #Divides(List< Expression >, Notation )
    */
   public /*constructor*/ Divides(List<Expression> elist) throws IllegalConstruction {
 	this(elist, null);
