+++ src/main/java/calculator/operations/Minus.java
@@ -1,4 +1,7 @@
-package calculator;
+package calculator.operations;
+
+import calculator.*;
+import calculator.numbers.RationalNumber;
 
 import java.util.List;
 
@@ -18,7 +21,7 @@ public final class Minus extends Operation
    *
    * @param elist The list of Expressions to subtract
    * @throws IllegalConstruction    If an empty list of expressions if passed as parameter
-   * @see #Minus(List<Expression>,Notation)
+   * @see #Minus(List< Expression >, Notation )
    */
   public /*constructor*/ Minus(List<Expression> elist) throws IllegalConstruction {
   	this(elist, null);
