+++ src/main/java/calculator/operations/Times.java
@@ -1,4 +1,7 @@
-package calculator;
+package calculator.operations;
+
+import calculator.*;
+import calculator.numbers.RationalNumber;
 
 import java.util.List;
 
@@ -17,7 +20,7 @@ public final class Times extends Operation
    *
    * @param elist The list of Expressions to multiply
    * @throws IllegalConstruction    If an empty list of expressions if passed as parameter
-   * @see #Times(List<Expression>,Notation)
+   * @see #Times(List< Expression >, Notation )
    */
   public /*constructor*/ Times(List<Expression> elist) throws IllegalConstruction {
   	this(elist, null);
