+++ src/test/java/calculator/TestAngleConverter.java
@@ -1,5 +1,6 @@
 package calculator;
 
+import calculator.numbers.AngleConverter;
 import org.junit.jupiter.api.Test;
 import static org.junit.jupiter.api.Assertions.*;
