+++ src/test/java/calculator/TestCounting.java
@@ -1,6 +1,11 @@
 package calculator;
 
 //Import Junit5 libraries for unit testing:
+import calculator.numbers.MyNumber;
+import calculator.operations.Divides;
+import calculator.operations.Minus;
+import calculator.operations.Plus;
+import calculator.operations.Times;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
