+++ src/test/java/calculator/TestEvaluator.java
@@ -2,6 +2,12 @@ package calculator;
 
 //Import Junit5 libraries for unit testing:
 import static org.junit.jupiter.api.Assertions.*;
+
+import calculator.numbers.MyNumber;
+import calculator.operations.Divides;
+import calculator.operations.Minus;
+import calculator.operations.Plus;
+import calculator.operations.Times;
 import org.junit.jupiter.api.*;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
