+++ src/test/java/calculator/TestMinus.java
@@ -2,6 +2,10 @@ package calculator;
 
 //Import Junit5 libraries for unit testing:
 import static org.junit.jupiter.api.Assertions.*;
+
+import calculator.numbers.MyNumber;
+import calculator.operations.Minus;
+import calculator.operations.Times;
 import org.junit.jupiter.api.*;
 
 import java.util.ArrayList;
