+++ src/test/java/calculator/TestOperation.java
@@ -2,6 +2,12 @@ package calculator;
 
 //Import Junit5 libraries for unit testing:
 import static org.junit.jupiter.api.Assertions.*;
+
+import calculator.numbers.MyNumber;
+import calculator.operations.Divides;
+import calculator.operations.Minus;
+import calculator.operations.Operation;
+import calculator.operations.Plus;
 import org.junit.jupiter.api.*;
 
 import java.util.Arrays;
