+++ src/test/java/calculator/TestRationalNumber.java
@@ -1,5 +1,6 @@
 package calculator;
 
+import calculator.numbers.RationalNumber;
 import org.junit.jupiter.api.Test;
 import static org.junit.jupiter.api.Assertions.*;
