+++ src/test/java/calculator/TestRealNumber.java
@@ -1,5 +1,6 @@
 package calculator;
 
+import calculator.numbers.RealNumber;
 import org.junit.jupiter.api.Test;
 import static org.junit.jupiter.api.Assertions.*;
