+++ src/test/java/calculator/TestRealOperations.java
@@ -1,5 +1,11 @@
 package calculator;
 
+import calculator.numbers.MyNumber;
+import calculator.numbers.RealNumber;
+import calculator.operations.Divides;
+import calculator.operations.Minus;
+import calculator.operations.Plus;
+import calculator.operations.Times;
 import org.junit.jupiter.api.Test;
 import static org.junit.jupiter.api.Assertions.*;
